requirements.txt
matplotlib==3.5.2
pandas==1.4.4
Pillow==9.3.0
streamlit==1.15.1
torch==1.13.0
torchvision==0.14.0
opencv_python_headless==4.6.0.66
IPython==7.31.1

